requirements.txt
flask
flask-cors
torch
transformers
numpy
google-genai
